feat: allow passing elements in proxy body
